requirements.txt
streamlit==1.32.0
langchain==0.1.12
groq==0.4.2
python-docx==1.1.0
PyMuPDF==1.23.26
pdfplumber==0.10.3
python-dotenv==1.0.1